requirements.txt
deepgram_sdk==3.1.0
langchain==0.1.0
langchain_community==0.0.13
langchain_core==0.1.11
langchain_google_genai==0.0.6
langchain_openai==0.0.2.post1
reportlab==3.6.8
streamlit==1.29.0
pinecone-client